implemented resume iteration for kmeans and minibatch kmeans.
Fixes #831

+++ pyemma/coordinates/clustering/kmeans.py
@@ -47,7 +47,7 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
 
     def __init__(self, n_clusters, max_iter=5, metric='euclidean',
                  tolerance=1e-5, init_strategy='kmeans++', fixed_seed=False,
-                 oom_strategy='memmap', stride=1, n_jobs=None, skip=0):
+                 oom_strategy='memmap', stride=1, n_jobs=None, skip=0, clustercenters=None, keep_data=False):
         r"""Kmeans clustering
 
         Parameters
@@ -91,15 +91,28 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
             Number of threads to use during assignment of the data.
             If None, all available CPUs will be used.
 
+        clustercenters: None or array(k, dim)
+            This is used to resume the kmeans iteration. Note, that if this is set, the init_strategy is ignored and
+            the centers are directly passed to the kmeans iteration algorithm.
+
+        keep_data: boolean, default False
+            If you intend to resume the kmeans iteration later on, in case it did not converge,
+            this parameter controls whether the input data is kept in memory or not.
+
         """
         super(KmeansClustering, self).__init__(metric=metric, n_jobs=n_jobs)
 
+        if clustercenters is None:
+            clustercenters = []
+
+        self._in_memory_chunks_set = False
+        self._converged = False
+
         self.set_params(n_clusters=n_clusters, max_iter=max_iter, tolerance=tolerance,
                         init_strategy=init_strategy, oom_strategy=oom_strategy,
-                        fixed_seed=fixed_seed, stride=stride, skip=skip
+                        fixed_seed=fixed_seed, stride=stride, skip=skip, clustercenters=clustercenters,
+                        keep_data=keep_data
                         )
-        # TODO: this could be an input parameter for estimation to resume the kmeans iteration.
-        self.clustercenters = []
 
     @property
     def init_strategy(self):
@@ -138,7 +151,22 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
 
         self.logger.debug("seed = %i", self._fixed_seed)
 
+    @property
+    def converged(self):
+        return self._converged
+
     def _init_in_memory_chunks(self, size):
+        # check if we need to allocate memory.
+        if hasattr(self, '_in_memory_chunks') and self._in_memory_chunks.size == size:
+            assert hasattr(self, '_in_memory_chunks')
+            self.logger.info("re-use in memory data.")
+            return
+        elif self._check_resume_iteration() and not self._in_memory_chunks_set:
+            pass
+            #self.logger.warning('Resuming kmeans iteration without the setting "keep_data=True", will re-create'
+            #                    ' the linear in-memory data. This is inefficient! Consider setting keep_data=True,'
+            #                    ' when you intend to resume.')
+
         available_mem = psutil.virtual_memory().available
         required_mem = self._calculate_required_memory(size)
         if required_mem <= available_mem:
@@ -167,25 +195,36 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
     def describe(self):
         return "[Kmeans, k=%i, inp_dim=%i]" % (self.n_clusters, self.data_producer.dimension())
 
-    def kmeanspp_center_assigned(self):
-        self._progress_update(1, stage=0)
+    def _check_resume_iteration(self):
+        # if we have centers set, we should continue to iterate these
+        return self.clustercenters.size != 0
 
     def _estimate(self, iterable, **kw):
         self._init_estimate()
 
-        with iterable.iterator(return_trajindex=True, stride=self.stride,
-                               chunk=self.chunksize, skip=self.skip) as iter:
-            # first pass: gather data and run k-means
-            first_chunk = True
-            for itraj, X in iter:
-                # collect data
-                self._collect_data(X, first_chunk)
-                # initialize cluster centers
-                self._initialize_centers(X, itraj, iter.pos, iter.last_chunk)
-                first_chunk = False
+        # collect the data only if, we have not done this previously (eg. keep_data=True) or the centers are not initialized.
+        if not self._check_resume_iteration() or not self._in_memory_chunks_set:
+            resume_centers = self._check_resume_iteration()
+            with iterable.iterator(return_trajindex=True, stride=self.stride,
+                                   chunk=self.chunksize, skip=self.skip) as iter:
+                # first pass: gather data and run k-means
+                first_chunk = True
+                for itraj, X in iter:
+                    # collect data
+                    self._collect_data(X, first_chunk, iter.last_chunk)
+                    if not resume_centers:
+                        # initialize cluster centers
+                        self._initialize_centers(X, itraj, iter.pos, iter.last_chunk)
+                    first_chunk = False
+            self._logger.debug("Accumulated all data, running kmeans on %s", self._in_memory_chunks.shape)
+            self._in_memory_chunks_set = True
+        else:
+            if len(self.clustercenters) != self.n_clusters:
+                # TODO: this can be non-fatal, because the extension can handle it?!
+                raise RuntimeError('Passed clustercenters do not match n_clusters: {} vs. {}'.
+                                   format(len(self.clustercenters), self.n_clusters))
 
         # run k-means with all the data
-        self._logger.debug("Accumulated all data, running kmeans on %s", self._in_memory_chunks.shape)
         it = 0
         converged_in_max_iter = False
         prev_cost = 0
@@ -196,29 +235,31 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
             prev_cost = cost
 
             if rel_change <= self.tolerance:
-                converged_in_max_iter = True
+                self._converged = True
                 self._logger.info("Cluster centers converged after %i steps.", it + 1)
                 self._progress_force_finish(stage=1)
                 break
             else:
                 self._progress_update(1, stage=1)
             it += 1
-        if not converged_in_max_iter:
+        if not self._converged:
             self._logger.info("Algorithm did not reach convergence criterion"
                               " of %g in %i iterations. Consider increasing max_iter.",
                               self.tolerance, self.max_iter)
-        # set centers
         self._finish_estimate()
 
         return self
 
     def _finish_estimate(self):
-        fh = None
-        if isinstance(self._in_memory_chunks, np.memmap):
-            fh = self._in_memory_chunks.filename
-        del self._in_memory_chunks
-        if fh:
-            os.unlink(fh)
+        # delete the large input array, if the user wants to keep the array or the estimate has converged.
+        if not self.keep_data or self._converged:
+            fh = None
+            if isinstance(self._in_memory_chunks, np.memmap):
+                fh = self._in_memory_chunks.filename
+            del self._in_memory_chunks
+            if fh:
+                os.unlink(fh)
+            self._in_memory_chunks_set = False
         if self.init_strategy == 'uniform':
             del self._init_centers_indices
         if self.init_strategy == 'kmeans++':
@@ -235,17 +276,15 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
         total_length = sum(traj_lengths)
         if not self.n_clusters:
             self.n_clusters = min(int(math.sqrt(total_length)), 5000)
-            assert isinstance(self.n_clusters, int)
             self._logger.info("The number of cluster centers was not specified, "
                               "using min(sqrt(N), 5000)=%s as n_clusters." % self.n_clusters)
-        # TODO: this should not be needed.
-        self.n_clusters = int(self.n_clusters)
         if self.init_strategy == 'kmeans++':
             self._progress_register(self.n_clusters,
                                     description="initialize kmeans++ centers", stage=0)
         self._progress_register(self.max_iter, description="kmeans iterations", stage=1)
         self._init_in_memory_chunks(total_length)
-        if not isinstance(self, MiniBatchKmeansClustering):
+        from pyemma.coordinates.data import DataInMemory
+        if not isinstance(self, MiniBatchKmeansClustering) and not isinstance(self.data_producer, DataInMemory):
             n_chunks = self.data_producer.n_chunks(chunksize=self.chunksize, skip=self.skip, stride=self.stride)
             self._progress_register(n_chunks, description="creating data array", stage='data')
 
@@ -258,7 +297,9 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
                             math.ceil((traj_len / float(total_length)) * self.n_clusters)))
 
         from ._ext import kmeans as kmeans_mod
-        self._inst = kmeans_mod.Kmeans_f(self.n_clusters, self.metric, self.data_producer.ndim, None)
+        self._inst = kmeans_mod.Kmeans_f(self.n_clusters, self.metric, self.data_producer.ndim,
+                                         (lambda: self._progress_update(1, stage=0))
+                                         if self.init_strategy == 'kmeans++' else None)
 
         return stride
 
@@ -274,10 +315,9 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
                         new = np.vstack((self.clustercenters, X[l]))
                         self.clustercenters = new
         elif last_chunk and self.init_strategy == 'kmeans++':
-            self._inst.set_callback(self.kmeanspp_center_assigned)
             self.clustercenters = self._inst.init_centers_KMpp(self._in_memory_chunks, self.fixed_seed)
 
-    def _collect_data(self, X, first_chunk):
+    def _collect_data(self, X, first_chunk, last_chunk):
         # beginning - compute
         if first_chunk:
             self._t_total = 0
@@ -285,24 +325,32 @@ class KmeansClustering(AbstractClustering, ProgressReporter):
         # appends a true copy
         self._in_memory_chunks[self._t_total:self._t_total + len(X)] = X[:]
         self._t_total += len(X)
-        if not isinstance(self, MiniBatchKmeansClustering):
+        from pyemma.coordinates.data import DataInMemory
+        if not isinstance(self, MiniBatchKmeansClustering) and not isinstance(self.data_producer, DataInMemory):
             self._progress_update(1, stage='data')
+
+        if last_chunk:
+            self._in_memory_chunks_set = True
 
 
 class MiniBatchKmeansClustering(KmeansClustering):
     r"""Mini-batch k-means clustering"""
 
     def __init__(self, n_clusters, max_iter=5, metric='euclidean', tolerance=1e-5, init_strategy='kmeans++',
-                 batch_size=0.2, oom_strategy='memmap', fixed_seed=False, stride=None, n_jobs=None, skip=0):
+                 batch_size=0.2, oom_strategy='memmap', fixed_seed=False, stride=None, n_jobs=None, skip=0,
+                 clustercenters=None, keep_data=False):
 
         if stride is not None:
             raise ValueError("stride is a dummy value in MiniBatch Kmeans")
         if batch_size > 1:
             raise ValueError("batch_size should be less or equal to 1, but was %s" % batch_size)
+        if keep_data:
+            raise ValueError("keep_data is a dummy value in MiniBatch Kmeans")
 
         super(MiniBatchKmeansClustering, self).__init__(n_clusters, max_iter, metric,
                                                         tolerance, init_strategy, False,
-                                                        oom_strategy, stride=stride, n_jobs=n_jobs, skip=skip)
+                                                        oom_strategy, stride=stride, n_jobs=n_jobs, skip=skip,
+                                                        clustercenters=clustercenters, keep_data=False)
 
         self.set_params(batch_size=batch_size)
 
@@ -346,12 +394,11 @@ class MiniBatchKmeansClustering(KmeansClustering):
         self._init_estimate()
 
         i_pass = 0
-        converged_in_max_iter = False
         prev_cost = 0
 
         ra_stride = self._draw_mini_batch_sample()
         with iterable.iterator(return_trajindex=False, stride=ra_stride, skip=self.skip) as iterator:
-            while not (converged_in_max_iter or i_pass + 1 >= self.max_iter):
+            while not (self._converged or i_pass + 1 > self.max_iter):
                 first_chunk = True
                 # draw new sample and re-use existing iterator instance.
                 ra_stride = self._draw_mini_batch_sample()
@@ -359,9 +406,9 @@ class MiniBatchKmeansClustering(KmeansClustering):
                 iterator.reset()
                 for X in iter(iterator):
                     # collect data
-                    self._collect_data(X, first_chunk)
+                    self._collect_data(X, first_chunk, iterator.last_chunk)
                     # initialize cluster centers
-                    if i_pass == 0:
+                    if i_pass == 0 and not self._check_resume_iteration():
                         self._initialize_centers(X, iterator.current_trajindex, iterator.pos, iterator.last_chunk)
                     first_chunk = False
 
@@ -369,11 +416,11 @@ class MiniBatchKmeansClustering(KmeansClustering):
                 self.clustercenters = self._inst.cluster(self._in_memory_chunks, self.clustercenters)
                 cost = self._inst.cost_function(self._in_memory_chunks, self.clustercenters)
 
-                rel_change = np.abs(cost - prev_cost) / cost
+                rel_change = np.abs(cost - prev_cost) / cost if cost != 0.0 else 0.0
                 prev_cost = cost
 
                 if rel_change <= self.tolerance:
-                    converged_in_max_iter = True
+                    self._converged = True
                     self._logger.info("Cluster centers converged after %i steps.", i_pass + 1)
                     self._progress_force_finish(stage=1)
                 else:
@@ -381,7 +428,9 @@ class MiniBatchKmeansClustering(KmeansClustering):
 
                 i_pass += 1
 
-        if not converged_in_max_iter:
+        self._finish_estimate()
+
+        if not self._converged:
             self._logger.info("Algorithm did not reach convergence criterion"
                               " of %g in %i iterations. Consider increasing max_iter.", self.tolerance, self.max_iter)
         return self
